+++ recipe-service/src/main/java/br/com/mateusulrich/recipeservice/ingredient/repository/IngredientRepository.java
@@ -10,10 +10,10 @@ import org.springframework.stereotype.Repository;
 import java.util.Optional;
 
 @Repository
-public interface IngredientRepository extends JpaRepository<Ingredient, Long>, JpaSpecificationExecutor<Ingredient> {
+public interface IngredientRepository extends JpaRepository<Ingredient, Integer>, JpaSpecificationExecutor<Ingredient> {
 
     @Query("select i from Ingredient i left join fetch i.possibleUnits where i.id = :id")
-    Optional<Ingredient> findIngredientWithUnits(@Param("id") Long id);
+    Optional<Ingredient> findIngredientWithUnits(@Param("id") Integer id);
 
     boolean existsByName(@Param("name") String name);
 }
